ml-service/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
supabase
python-dotenv
numpy
pydantic
